Refactor the do-nothing DatasetGraphs.
